pyproject.toml
[tool.poetry]
name = "langfuse-trial"
version = "0.1.0"
description = ""
authors = ["Kogia-sima <orcinus4627@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
main = "langfuse_trial.main:main"

[tool.poetry.dependencies]
python = "^3.11"
langfuse = "^2.40.0"
langchain = "^0.2.12"
langchain-community = "^0.2.11"
python-dotenv = "^1.0.1"
openai = "^1.38.0"
tiktoken = "^0.7.0"
faiss-cpu = "^1.8.0.post1"
langchain-openai = "^0.1.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
